fix chart download
