+++ web/src/features/responses/components/AggregateCard/AggregateCard.tsx
@@ -1,0 +1,69 @@
+import { Square2StackIcon } from '@heroicons/react/24/solid';
+import { useRef } from 'react';
+import type { FunctionComponent } from '@/common/types';
+import { saveChart } from '@/components/charts/utils/save-chart';
+import { Button } from '@/components/ui/button';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
+import {
+  type BaseQuestionAggregate,
+  isDateAggregate,
+  isMultiSelectAggregate,
+  isNumberAggregate,
+  isRatingAggregate,
+  isSingleSelectAggregate,
+  isTextAggregate,
+} from '../../models/form-aggregated';
+import { DateAggregateContent } from '../DateAggregateContent/DateAggregateContent';
+import { MultiSelectAggregateContent } from '../MultiSelectAggregateContent/MultiSelectAggregateContent';
+import { NumberAggregateContent } from '../NumberAggregateContent/NumberAggregateContent';
+import { RatingAggregateContent } from '../RatingAggregateContent/RatingAggregateContent';
+import { SingleSelectAggregateContent } from '../SingleSelectAggregateContent/SingleSelectAggregateContent';
+import { TextAggregateContent } from '../TextAggregateContent/TextAggregateContent';
+
+type AggregateCardProps = {
+  aggregate: BaseQuestionAggregate;
+  language: string;
+};
+
+export function AggregateCard({ aggregate, language }: AggregateCardProps): FunctionComponent {
+  const chartRef = useRef(null);
+
+  return (
+    <Card key={aggregate.questionId} className='max-w-4xl'>
+      <CardHeader>
+        <CardTitle className='text-xl flex justify-between items-center'>
+          {aggregate.question.code}. {aggregate.question.text[language]}
+          {!isTextAggregate(aggregate) && !isRatingAggregate(aggregate) && (
+            <Button
+              className='gap-1'
+              onClick={() => {
+                saveChart(chartRef, '');
+              }}
+              variant='outline'>
+              <Square2StackIcon className='h-4 w-4' /> Copy chart
+            </Button>
+          )}
+        </CardTitle>
+        <p className='text-gray-500'>{aggregate.answersAggregated} answers</p>
+      </CardHeader>
+
+      <CardContent>
+        {isDateAggregate(aggregate) && <DateAggregateContent ref={chartRef} aggregate={aggregate} />}
+
+        {isMultiSelectAggregate(aggregate) && (
+          <MultiSelectAggregateContent ref={chartRef} aggregate={aggregate} language={language} />
+        )}
+
+        {isNumberAggregate(aggregate) && <NumberAggregateContent ref={chartRef} aggregate={aggregate} />}
+
+        {isRatingAggregate(aggregate) && <RatingAggregateContent aggregate={aggregate} />}
+
+        {isSingleSelectAggregate(aggregate) && (
+          <SingleSelectAggregateContent ref={chartRef} aggregate={aggregate} language={language} />
+        )}
+
+        {isTextAggregate(aggregate) && <TextAggregateContent aggregate={aggregate} />}
+      </CardContent>
+    </Card>
+  );
+}
